Use priority queue instead of sort in multiplication
This doesn't really change the asymptotic complexity, but it is a
version of the 'diagonal trick' (e76621fea) that works with *any*
monomial ordering. In practise, this implementation is slightly faster
even for degrevlex (unfortunately, for now, not so much faster as the
time I spent on this should warrant :( ).

Priority queue with fewer allocations

Remove sort based multiplication